Fix the DiskUsageBar.
diskUsageBar.contentWidth is undefined since the variable was moved
into the update function so it's only calculated when we need to.

Only update totalspace after 200ms. It was constantly calling onChanged
whenever usedspace and freespace were out of sync, which triggered a
UI update.

Ditch the RowLayout stuff and manually calculate the partition
minWidth and width using a simple Row.