Remove transparent fill and bold stroked line.
Transparent fill areas are too small usually for the distiction of a
bold stroked line to matter. The CPU Graph is already cluttered enough
and the buffered RAM is usually thin too. Tiny uploads are hard to
tell the difference between the pink/purple now that there's 4 colors,
where the 2 extra colors are the computed colors calculated using the
colors underneath.